refactor(oracle): migrate bgp to per-protocol plugins
