+++ tools/oracle/engine/generator.py
@@ -66,6 +66,12 @@ from .protocols import SAMPLER_REGISTRY
 # still call them to map an IGMP case onto its feature/behavior (the same
 # co-locate-and-re-import pattern as the IPv6 ext-header builders in ``packets``).
 from .protocols.igmp import _igmp_behavior_for_case, _igmp_feature_for_case
+# ``_apply_bgp_behavior`` moved to the BGP sampler plugin (``protocols/bgp.py``)
+# with the rest of the BGP feature behavior. It is re-imported here because the
+# focused bgp-smoke profile path in ``generate_plan`` calls it directly (with
+# ``behavior=""``) outside the registry feature loop, the same co-locate-and-
+# re-import pattern as the IGMP case/behavior helpers above.
+from .protocols.bgp import _apply_bgp_behavior
 
 
 SUPPORTED_LAYER_BACKEND = "libcrafter"
@@ -239,7 +245,6 @@ _SUPPORTED_FIELDS: dict[str, set[str]] = {
         "iv",
         "icv",
     },
-    "bgp": {"marker", "length", "type"},
     "rip": {"command", "version", "reserved"},
     "ripng": {"command", "version", "reserved"},
     "ikev2": {
@@ -1686,9 +1691,7 @@ class PacketGenerator:
                     grammar=self.grammar,
                 )
                 return
-        if feature is not None and feature.startswith("bgp_") and "bgp" in fields:
-            _apply_bgp_behavior(fields, stack=stack, case=case, behavior=behavior)
-        elif feature is not None and feature.startswith("ripng_") and "ripng" in fields:
+        if feature is not None and feature.startswith("ripng_") and "ripng" in fields:
             _apply_ripng_behavior(fields, stack=stack, case=case, behavior=behavior)
         elif feature is not None and feature.startswith("rip_") and "rip" in fields:
             _apply_rip_behavior(fields, stack=stack, case=case, behavior=behavior)
@@ -1825,8 +1828,6 @@ class PacketGenerator:
                 field_spec=field_spec,
                 current_fields=current_fields,
             )
-        if layer == "bgp":
-            return _sample_bgp_field(ctx, field_name, domain)
         if layer == "rip":
             return _sample_rip_field(ctx, field_name, domain)
         if layer == "ripng":
@@ -1999,16 +2000,6 @@ def _ipv6_extension_options_for_case(layer: str, case: str) -> list[JSONObject]:
         {"kind": "home_address", "address": "2001:db8::42"},
         {"kind": "padn", "total_length": 4},
     ]
-
-
-def _sample_bgp_field(ctx: _SamplingContext, field_name: str, domain: object) -> object:
-    if field_name == "marker":
-        return {"hex": "ff" * 16}
-    if field_name == "length":
-        return _SKIP_FIELD
-    if field_name == "type":
-        return _bgp_message_type_for_case(ctx.case)
-    raise ValueError(f"spec error: unsupported bgp field sampler: {field_name}")
 
 
 def _sample_rip_field(ctx: _SamplingContext, field_name: str, domain: object) -> object:
@@ -3673,43 +3664,6 @@ def _dns_answers_for_domain(ctx: _SamplingContext, domain: object) -> list[JSONO
     return [{"name": "example.com.", "type": "A", "ttl": 60, "address": ctx.dst_ipv4}]
 
 
-def _apply_bgp_behavior(
-    fields: dict[str, JSONObject],
-    *,
-    stack: Sequence[str],
-    case: str,
-    behavior: str,
-) -> None:
-    bgp = fields["bgp"]
-    bgp["marker"] = {"hex": "ff" * 16}
-    bgp["type"] = _bgp_message_type_for_case(case)
-    bgp["message_type"] = bgp["type"]
-
-    if "tcp" in fields:
-        fields["tcp"]["dst_port"] = 179
-    if "payload" in fields:
-        fields["payload"] = {"hex": "", "length": 0}
-
-    if bgp["type"] == "keepalive":
-        bgp.pop("body", None)
-        return
-
-    if bgp["type"] == "open":
-        bgp["body"] = {"hex": _bgp_open_body_hex(case)}
-        return
-
-    if bgp["type"] == "notification":
-        bgp["body"] = {"hex": "0203deadbeef"}
-        return
-
-    if bgp["type"] == "route_refresh":
-        bgp["body"] = {"hex": "00010001"}
-        return
-
-    if bgp["type"] == "update":
-        bgp["body"] = {"hex": _bgp_update_body_hex(stack=stack, case=case, behavior=behavior)}
-
-
 def _apply_ospf_behavior(
     fields: dict[str, JSONObject],
     *,
@@ -3835,61 +3789,6 @@ def _is_ospf_smoke_case(case: str) -> bool:
     """Whether ``case`` is sampled by the focused ospf-smoke profile."""
 
     return case.replace("_", "-").startswith("ospf-")
-
-
-def _bgp_message_type_for_case(case: str) -> str:
-    normalized = case.replace("_", "-")
-    if "keepalive" in normalized:
-        return "keepalive"
-    if "notification" in normalized:
-        return "notification"
-    if "route-refresh" in normalized:
-        return "route_refresh"
-    if "open" in normalized:
-        return "open"
-    return "update"
-
-
-def _bgp_open_body_hex(case: str) -> str:
-    base = "04fc00005ac0000201"
-    if "capabilities" not in case.replace("_", "-"):
-        return base + "00"
-    capabilities = (
-        "010400010001"  # Multiprotocol IPv4 unicast.
-        "010400020001"  # Multiprotocol IPv6 unicast.
-        "41040000fc00"  # Four-octet AS 64512.
-        "0200"          # Route refresh.
-    )
-    optional_parameters = "02" + f"{len(bytes.fromhex(capabilities)):02x}" + capabilities
-    return base + f"{len(bytes.fromhex(optional_parameters)):02x}" + optional_parameters
-
-
-def _bgp_update_body_hex(*, stack: Sequence[str], case: str, behavior: str) -> str:
-    normalized = f"{case} {behavior}".replace("_", "-")
-    if "withdraw" in normalized:
-        return "000418c000020000"
-    if "mp-reach" in normalized:
-        mp_reach_value = "00020110" + "20010db8000000000000000000000001" + "00" + "2020010db8"
-        attrs = "800e" + f"{len(bytes.fromhex(mp_reach_value)):02x}" + mp_reach_value
-        return "0000" + len(bytes.fromhex(attrs)).to_bytes(2, "big").hex() + attrs
-    if "extended-communities" in normalized:
-        attrs = "c01010" + "0002fc0000000064" + "0002fc00000000c8"
-        return "0000" + len(bytes.fromhex(attrs)).to_bytes(2, "big").hex() + attrs
-    if "large-communities" in normalized:
-        attrs = (
-            "c02018"
-            "0000fc000000006400000001"
-            "0000fc000000006400000002"
-        )
-        return "0000" + len(bytes.fromhex(attrs)).to_bytes(2, "big").hex() + attrs
-    if "communities" in normalized:
-        attrs = "c00808ffffff0100fc0064"
-        return "0000" + len(bytes.fromhex(attrs)).to_bytes(2, "big").hex() + attrs
-    attrs = "40010100" + "4002040201fc00" + "400304c0000201"
-    nlri = "18c63364"
-    if "announce" not in normalized and "ipv6" in stack:
-        nlri = ""
-    return "0000" + len(bytes.fromhex(attrs)).to_bytes(2, "big").hex() + attrs + nlri
 
 
 # --------------------------------------------------------------------------
