+++ tools/oracle/tests/test_protocol_plugin_coverage.py
@@ -41,7 +41,6 @@ from tools.oracle.engine.spec_loader import load_oracle_specs
 KNOWN_UNMIGRATED: frozenset[str] = frozenset(
     {
         "ah",
-        "bgp",
         "ble_adv",
         "ble_radio",
         "dhcp",
